+++ crazykashark/src/main/java/com/kashark/crazykashark/entity/PlanGroup.java
@@ -12,19 +12,19 @@ import java.sql.Timestamp;
  */
 @Data
 public class PlanGroup {
-    private Long id;
+    private Integer id;
 
-    private Long studentId;
+    private Integer studentId;
 
     private String coreTeacherId;
 
     private Timestamp createTime;
 
-    private Long createBy;
+    private Integer createBy;
 
     private Timestamp updateTime;
 
-    private Long updateBy;
+    private Integer updateBy;
 
     @TableLogic
     private Integer deleted;
