+++ crazykashark/src/main/java/com/kashark/crazykashark/entity/Staff.java
@@ -11,9 +11,9 @@ import java.sql.Timestamp;
  */
 @Data
 public class Staff {
-    private Long id;
+    private Integer id;
 
-    private Long staffId;
+    private Integer staffId;
 
     private String staffName;
 
@@ -29,11 +29,11 @@ public class Staff {
 
     private Timestamp createTime;
 
-    private Long createBy;
+    private Integer createBy;
 
     private Timestamp updateTime;
 
-    private Long updateBy;
+    private Integer updateBy;
 
     private Integer deleted;
 }
